+++ src/main/java/duke/Ui.java
@@ -1,3 +1,8 @@
+package duke;
+
+import duke.task.Task;
+import duke.task.TaskList;
+
 public class Ui {
     private static final String logo =
             "$$\\                                     $$\\\n" +
